+++ src/engine/server/common.rs
@@ -21,7 +21,7 @@ pub enum BlockType {
 #[derive(Debug, Clone, Copy)]
 pub struct BlockArray {
     pub block_type: [BlockType; CHUNK_BLOCK_COUNT as usize],
-    pub block_id: [u16; CHUNK_BLOCK_COUNT as usize],
+    pub block_id: [u32; CHUNK_BLOCK_COUNT as usize],
     pub texture_index: [u8; CHUNK_BLOCK_COUNT as usize],
 }
 
@@ -34,11 +34,11 @@ impl BlockArray {
         self.block_type[(chunk_relative_pos.y * CHUNK_SIZE + chunk_relative_pos.x) as usize] = block_type;
     }
 
-    pub fn set_block_id_byindex(&mut self, index: usize, id: u16) {
+    pub fn set_block_id_byindex(&mut self, index: usize, id: u32) {
         self.block_id[index] = id;
     }
 
-    pub fn set_block_id(&mut self, chunk_relative_pos: ChunkRelativePos, id: u16) {
+    pub fn set_block_id(&mut self, chunk_relative_pos: ChunkRelativePos, id: u32) {
         self.block_id[(chunk_relative_pos.y * CHUNK_SIZE + chunk_relative_pos.x) as usize] = id;
     }
 
@@ -57,7 +57,7 @@ impl BlockArray {
     }
 
     pub fn clear_block(&mut self, chunk_relative_pos: ChunkRelativePos) {
-        let pos = chunk_relative_pos.y * CHUNK_SIZE + chunk_relative_pos.x;
+        let pos = chunk_relative_pos.y as usize * CHUNK_SIZE as usize + chunk_relative_pos.x as usize;
         self.block_type[pos as usize] = BlockType::Air;
         self.block_id[pos as usize] = 0;
         self.texture_index[pos as usize] = 0;
